Standardize HP2's TEXF naming based on UT469's convensions
No reason to not do that, since 1. BC6 isn't implemented yet in HP2 (so the actual end format is up in the air) and 2. BGRA is already what the raw 32bit color texture type is. I just hadn't gotten around to renaming it throughout the whole solution. I have #ifdef'd it for all other games though (back to TEXF_RGBA8)